requirements.txt
requests
uuid
cchardet